Refactor: Move LLLnode into it's own file.
